backend/pyproject.toml
[tool.poetry]
name = "hr-monitor-6146"
version = "0.1.0"
description = ""
authors = ["Александр Мамедов <anifan01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
sqlalchemy = "^2.0.35"
ruff = "^0.6.9"
uvicorn = "^0.31.0"
alembic = "^1.13.3"
python-jose = "^3.3.0"
pydantic-settings = "^2.5.2"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
greenlet = "^3.1.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
